requirements.txt
altgraph==0.17.2
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
importlib-metadata==4.8.2
macholib==1.15.2
Pillow==8.4.0
pycodestyle==2.8.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
qrcode==7.3.1
requests==2.26.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
